Finalized motor/sensor data converters

+++ bluetoothsdk/src/main/java/org/revolutionrobotics/robotcontroller/bluetooth/service/MotorInfo.kt
@@ -1,3 +1,3 @@
 package org.revolutionrobotics.robotcontroller.bluetooth.service
 
-data class MotorInfo(val speed: Float, val position: Int, val power: Long)
+data class MotorInfo(val speed: Float, val position: Int, val power: Int)
