+++ bluetoothsdk/src/main/java/org/revolutionrobotics/robotcontroller/bluetooth/service/UltrasoundInfo.kt
@@ -1,3 +1,3 @@
 package org.revolutionrobotics.robotcontroller.bluetooth.service
 
-data class UltrasoundInfo(val distance: Long)
+data class UltrasoundInfo(val distance: Int)
